+++ noodles-gff/src/io/writer/line/record.rs
@@ -21,7 +21,7 @@ where
     W: Write,
     R: Record + ?Sized,
 {
-    write_reference_sequence_name(writer, record.reference_sequence_name())?;
+    write_reference_sequence_name(writer, record.reference_sequence_name().as_ref())?;
 
     write_separator(writer)?;
     write_source(writer, record.source())?;