+++ noodles-gtf/src/io/writer/line/record.rs
@@ -22,7 +22,7 @@ where
     W: Write,
     R: Record + ?Sized,
 {
-    write_reference_sequence_name(writer, record.reference_sequence_name())?;
+    write_reference_sequence_name(writer, record.reference_sequence_name().as_ref())?;
 
     write_separator(writer)?;
     write_source(writer, record.source())?;