+++ noodles-gtf/src/record.rs
@@ -3,7 +3,7 @@
 pub mod attributes;
 mod fields;
 
-use std::{fmt, io};
+use std::{borrow::Cow, fmt, io};
 
 use bstr::{BStr, ByteSlice};
 use noodles_core::Position;
@@ -89,8 +89,8 @@ impl fmt::Debug for Record<'_> {
 }
 
 impl gff::feature::Record for Record<'_> {
-    fn reference_sequence_name(&self) -> &BStr {
-        self.reference_sequence_name()
+    fn reference_sequence_name(&self) -> Cow<'_, BStr> {
+        Cow::from(self.reference_sequence_name())
     }
 
     fn source(&self) -> &BStr {